添加ssh判断解析函数
